backend/requirements.txt
Django>=5.0,<6
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-environ
django-filter
